feat: swap releaser
This commit adds a new base releaser called the SwapReleaser which takes
a threshold of UNI and sends it to a configured address. This base
contract is then easily inherited by Firepit to set the burn address as
the recipient. This functionality enables the use the same competitive
release logic for alternate non-burn mechanisms in the future